Backend/pyproject.toml
[project]
name = "triplink-backend"
version = "0.1.0"
description = "TripLink Flask backend."
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.0.5",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "stripe>=13.0.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "pytest>=8.4.2",
]
